Add header and footer size options to the config
[html-pdf](https://www.npmjs.com/package/html-pdf) package allows the use of headers and footers which are placed automatically at the top and bottom of each page, a code sample is below.

```html
<div id="pageHeader">
  <center>Elliot Blackburn</center>
  <center>My document header</center>
</div>
```

The html-pdf default height for a header is 40mm which is a quite big, but the html-pdf package allows [configuration of this value](https://www.npmjs.com/package/html-pdf#options) in the options object.

As the the html-pdf options are currently read in directly from this packages configuration, I've added two new options with defaults for the user to modify.

This solution isn't perfect though, using a default of 0mm means when the user adds a header or footer for the first time without modifying these values, the header or footer won't display. This could be confusing for new years trying to make use of the feature.